Add HubHelpers module
